move field to top of main CborReader source file.
